requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
ffmpeg==1.4
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.0
jupyterlab-server==1.2.0
kiwisolver==1.2.0
m3u8==0.3.12
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
oauthlib==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-twitch-client==0.6.0
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
twitch-dl==1.9.0
txt2tags==3.7
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
zmq==0.0.0
